chore: extract `hooks` for FontPicker
